+++ packages/nextjs/components/Processes/BundlingProcess/Steps/AssetSelectionStep/AutoDetectedAssets/AutoDetectedAssetItem.tsx
@@ -1,0 +1,68 @@
+import Image from "next/image";
+import LogoSvg from "../../../../../../public/assets/flashbotRecovery/logo.svg";
+import styles from "./autoDetectedAssets.module.css";
+import { motion } from "framer-motion";
+import { RecoveryTx } from "~~/types/business";
+
+interface IAssetProps {
+  onClick: () => void;
+  isSelected: boolean;
+  tx?: RecoveryTx;
+  isLoading: boolean;
+}
+export const AutoDetectedAssetItem = ({ onClick, isSelected, tx, isLoading }: IAssetProps) => {
+  const getSubtitleTitle = () => {
+    if (!tx) {
+      return "";
+    }
+    if (tx.type == "erc721") {
+      //@ts-ignore
+      return `Token ID: ${tx.tokenId!}`;
+    }
+    if (tx.type == "erc1155") {
+      //@ts-ignore
+      return `Token IDs: ${tx.tokenIds.map(hexId => BigInt(hexId).toString())}`;
+    }
+    if (tx.type === "erc20") {
+      //@ts-ignore
+      return tx.value;
+    }
+    if (tx.type === "custom") {
+      //@ts-ignore
+      return tx.info.split(" to ")[1];
+    }
+    return "";
+  };
+
+  const getTitle = () => {
+    if (!tx) {
+      return "";
+    }
+    if (tx.type === "erc20") {
+      //@ts-ignore
+      return tx.symbol;
+    }
+    if (tx.type === "custom") {
+      //@ts-ignore
+      return tx.info.split(" to ")[0];
+    }
+    return tx.info;
+  };
+
+  return (
+    <motion.div
+      initial={{ opacity: 0 }}
+      animate={{ opacity: 1 }}
+      onClick={onClick}
+      className={`${isSelected ? "bg-base-200" : ""} ${styles.assetItem}  ${isLoading ? styles.loading : ""}`}
+    >
+      <div className={`${styles.logoContainer}`}>
+        <Image className={styles.logo} src={LogoSvg} alt="" />
+      </div>
+      <div className={`${styles.data}`}>
+        <h3>{getTitle()}</h3>
+        <span>{getSubtitleTitle()}</span>
+      </div>
+    </motion.div>
+  );
+};
