+++ packages/nextjs/components/Processes/BundlingProcess/Steps/AssetSelectionStep/ManualAssetSelection/BasicFlow/types.d.ts
@@ -1,0 +1,6 @@
+export interface ITokenForm {
+  hackedAddress: string;
+  safeAddress: string;
+  close: () => void;
+  addAsset: (arg: RecoveryTx) => void;
+}
