+++ packages/nextjs/components/Processes/BundlingProcess/Steps/AssetSelectionStep/assetSelectionStep.module.css
@@ -1,0 +1,17 @@
+.container {
+  width: 100%;
+
+  margin-bottom: 0;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.title {
+  font-size: 23px;
+  font-style: normal;
+  font-weight: 600;
+  line-height: 100%;
+  margin: 0;
+}
