--- packages/nextjs/components/flashbotRecovery/AssetSelectionStep/assetSelectionStep.module.css
@@ -1,78 +1,0 @@
-.container{
-    width: 100%;
-   
-    margin-bottom: 0;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-
-.title{
-    font-size: 23px;
-    font-style: normal;
-    font-weight: 600;
-    line-height: 100%;
-    margin: 0;
-}
-.assetItem{
-    display: flex;
-    height: 92px;
-    max-width: 458px;
-    margin: 0 auto;
-    border-radius: 12px;
-    cursor: pointer;
-    margin-bottom: 12px;
-}
-.data{
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-}
-.assetList{
-    width: 100%;
-    max-height: calc(100vh - 400px);
-    height: calc(100vh - 400px);
-    overflow: auto;
-    margin: 36px 0;
-}
-
-.logo{
-    width: 60px;
-    height: 60px;
-    margin-right: 16px;
-}
-
-.logoContainer{
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    margin-left:16px;
-}
-
-.loader{
-    height: 94px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.loading span{
-    background-color:#3a4862;
-    width: 100px;
-    height: 24px;
-}
-.loading h3{
-    background-color:#3a4862;
-    width: 130px;
-    height: 24px;
-}
-
-.loading{
-    animation: loading 1s infinite;
-}
-
-@keyframes loading {
-    from {opacity: 0.3;}
-    to {opacity: 0.8;}
-  }
