--- packages/nextjs/components/flashbotRecovery/Layout/layout.module.css
@@ -1,121 +1,0 @@
-.layout {
-  display: flex;
-  justify-content: flex-start;
-  align-items: flex-start;
-  width: 100%;
-  height: 100vh;
-  position: absolute;
-  top: 0;
-  left: 0;
-  bottom: 0;
-  right: 0;
-}
-.sidebar {
-  max-width: 530px;
-  display: flex;
-  width: auto;
-
-  background-color: #243148;
-  height: 100%;
-}
-
-.sidebarContent {
-  margin-top: 56px;
-  margin-left: 10px;
-}
-.content {
-  display: flex;
-  width: 100%;
-  height: 100%;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.logoContainer {
-  display: flex;
-  align-items: center;
-  font-size: 23px;
-  font-style: normal;
-  font-weight: 400;
-}
-
-.logo {
-  width: 60px;
-  height: 60px;
-  margin-right: 16px;
-}
-.steps {
-  margin-top: 48px;
-  margin-right: 63px;
-  position: relative;
-}
-.step {
-  display: flex;
-  margin-top: 16px;
-}
-.badge {
-  border-radius: 100px;
-  border: 2px solid;
-  display: flex;
-  width: 34px;
-  font-size: 20px;
-  font-style: normal;
-  font-weight: 600;
-  height: 34px;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-.stepContainer {
-  margin-left: 16px;
-}
-.stepTitle {
-  font-size: 18px;
-  font-style: normal;
-  font-weight: 600;
-  margin: 0;
-  display: flex;
-  height: 34px;
-  align-items: center;
-}
-.stepDescription {
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 26px;
-  margin: 0;
-}
-
-.addresess {
-  display: flex;
-  position: absolute;
-  bottom: 0;
-  width: 100%;
-  border-radius: 12px;
-  justify-content: space-around;
-  max-width: 380px;
-  width: 34%;
-  left: 96px;
-  bottom: 26px;
-}
-
-.addressContainer {
-  margin: 24px 0;
-  margin-right: 0;
-}
-.addressContainer svg {
-  display: none;
-}
-.completed{
-  color: #7E92BD;
-}
-.completed .badge{
-  background-color: #3A4A6E ;
-
-}
-@media (min-width: 768px) {
-  .sidebarContent {
-    margin-left: 96px;
-  }
-}
