--- packages/nextjs/components/flashbotRecovery/RecoveryProcess/RecoveryProcess.tsx
@@ -1,148 +1,0 @@
-import React from "react";
-import { CustomPortal } from "../CustomPortal/CustomPortal";
-import { RecoveryProcessStatus } from "~~/hooks/flashbotRecoveryBundle/useRecoveryProcess";
-import LogoSvg from "~~/public/assets/flashbotRecovery/logo.svg";
-import ClockSvg from "~~/public/assets/flashbotRecovery/clock.svg";
-import SuccessSvg from "~~/public/assets/flashbotRecovery/success.svg";
-import VideoSvg from "~~/public/assets/flashbotRecovery/video.svg";
-
-interface IProps {
-  recoveryStatus: RecoveryProcessStatus;
-  startSigning: () => void;
-  finishProcess: () => void;
-  startProcess: () => void;
-  connectedAddress: string;
-  safeAddress: string;
-  hackedAddress: string;
-  blockCountdown:number
-}
-export const RecoveryProcess = ({
-  recoveryStatus,
-  startSigning,
-  startProcess,
-  finishProcess,
-  blockCountdown,
-  connectedAddress,
-  safeAddress,
-  hackedAddress,
-}: IProps) => {
-
-  if (recoveryStatus == RecoveryProcessStatus.initial) {
-    return <></>;
-  }
-  if (recoveryStatus == RecoveryProcessStatus.gasCovered) {
-    alert("you already covered the gas. If you're in a confussy situation, clear cookies and refresh page.");
-    return;
-  }
-
-  if (recoveryStatus == RecoveryProcessStatus.cachedDataToClean) {
-    return (
-      <CustomPortal
-        title={"Clear cache"}
-        description={
-          "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nunc vulputate libero et velit interdum, ac aliquet odio mattis. Class aptent taciti."
-        }
-        image={VideoSvg}
-      />
-    );
-  }
-
-  if (recoveryStatus == RecoveryProcessStatus.noSafeAccountConnected) {
-    return (
-      <CustomPortal
-        title={"Switch to safe address"}
-        description={
-          "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nunc vulputate libero et velit interdum, ac aliquet odio mattis. Class aptent taciti."
-        }
-        button={{
-          text: "Continue",
-          disabled: connectedAddress !== safeAddress,
-          action: () => startProcess(),
-        }}
-        image={LogoSvg}
-      />
-    );
-  }
-  if (
-    recoveryStatus == RecoveryProcessStatus.switchFlashbotNetworkAndPayBundleGas
-  ) {
-    return (
-      <CustomPortal
-        title={"Switching Network"}
-        description={"Switch to personal Flashbot RPC network to prepare the transacion you will pay"}
-        image={VideoSvg}
-      />
-    );
-  }
-
-  if(recoveryStatus == RecoveryProcessStatus.increaseGas){
-    return (
-      <CustomPortal
-        title={"Increase the gas"}
-        description={"Switch to personal Flashbot RPC network to prepare the transacion you will pay"}
-        image={VideoSvg}
-      />
-    );
-  }
-  if (recoveryStatus == RecoveryProcessStatus.switchToHacked) {
-    return (
-      <CustomPortal
-        title={"Switch to hacked address"}
-        description={
-          "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nunc vulputate libero et velit interdum, ac aliquet odio mattis. Class aptent taciti."
-        }
-        button={{
-          text: "Continue",
-          disabled: connectedAddress !== hackedAddress,
-          action: () => startSigning(),
-        }}
-        image={LogoSvg}
-      />
-    );
-  }
-  if (
-    recoveryStatus == RecoveryProcessStatus.signEachTransaction ||
-    recoveryStatus == RecoveryProcessStatus.allTxSigned ||
-    recoveryStatus == RecoveryProcessStatus.sendingBundle
-  ) {
-    return (
-      <CustomPortal
-        title={"Sign each transaction"}
-        description={
-          "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nunc vulputate libero et velit interdum, ac aliquet odio mattis. Class aptent taciti."
-        }
-        image={LogoSvg}
-      />
-    );
-  }
-  if (recoveryStatus == RecoveryProcessStatus.listeningToBundle) {
-    return (
-      <CustomPortal
-        title={"Wait without refreshing the page"}
-        description={
-          "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nunc vulputate libero et velit interdum, ac aliquet odio matt"
-        }
-        image={ClockSvg}
-        indicator={blockCountdown}
-      />
-    );  
-  }
-  if (recoveryStatus == RecoveryProcessStatus.success) {
-    return (
-      <CustomPortal
-        title={"Your assets have been recovered!"}
-        description={
-          "If we have helped you please, share the tool and if you want let us a tip to continue contributing"
-        }
-        button={{
-          text: "Continue",
-          disabled: connectedAddress !== hackedAddress,
-          action: () => finishProcess(),
-        }}
-        image={SuccessSvg}
-      />
-    );
-  }
-
-  return <></>;
-};
