+++ packages/nextjs/pages/index.tsx
@@ -1,69 +1,68 @@
-/* eslint-disable */
-import { SetStateAction, useEffect, useState } from "react";
+import { useState } from "react";
 import dynamic from "next/dynamic";
-import LogoSvg from "../public/assets/flashbotRecovery/logo.svg";
-import VideoSvg from "../public/assets/flashbotRecovery/video.svg";
 import { BigNumber } from "ethers";
-import { isAddress } from "ethers/lib/utils";
 import { NextPage } from "next";
 import { useLocalStorage } from "usehooks-ts";
 import { useAccount } from "wagmi";
+import { CustomHeader } from "~~/components/CustomHeader/CustomHeader";
 import { MetaHeader } from "~~/components/MetaHeader";
-import { AssetSelectionStep } from "~~/components/flashbotRecovery/AssetSelectionStep/AssetSelectionStep";
-import { ConnectStep } from "~~/components/flashbotRecovery/ConnectStep/ConnectStep";
-import { CustomHeader } from "~~/components/flashbotRecovery/CustomHeader/CustomHeader";
-import { CustomPortal } from "~~/components/flashbotRecovery/CustomPortal/CustomPortal";
-import { HackedAddressStep } from "~~/components/flashbotRecovery/HackedAddressStep/HackedAddressStep";
-import { Layout } from "~~/components/flashbotRecovery/Layout/Layout";
-import { RecoveryProcess } from "~~/components/flashbotRecovery/RecoveryProcess/RecoveryProcess";
-import { TransactionBundleStep } from "~~/components/flashbotRecovery/TransactionBundleStep/transactionBundleStep";
-import { RecoveryProcessStatus, useRecoveryProcess } from "~~/hooks/flashbotRecoveryBundle/useRecoveryProcess";
+import { BundlingProcess } from "~~/components/Processes/BundlingProcess/BundlingProcess";
+import { ConnectionProcess } from "~~/components/Processes/ConnectionProcess/ConnectionProcess";
+import { RecoveryProcess } from "~~/components/Processes/RecoveryProcess/RecoveryProcess";
+import { useRecoveryProcess } from "~~/hooks/flashbotRecoveryBundle/useRecoveryProcess";
 import { RecoveryTx } from "~~/types/business";
+import { BundlingSteps, RecoveryProcessStatus } from "~~/types/enums";
 
 const Home: NextPage = () => {
-  const { isConnected, address } = useAccount();
+  const { isConnected: walletConnected, address: connectedAddress } = useAccount();
   const [safeAddress, setSafeAddress] = useLocalStorage<string>("toAddress", "");
   const [hackedAddress, setHackedAddress] = useLocalStorage<string>("hackedAddress", "");
   const [unsignedTxs, setUnsignedTxs] = useLocalStorage<RecoveryTx[]>("unsignedTxs", []);
   const [totalGasEstimate, setTotalGasEstimate] = useState<BigNumber>(BigNumber.from("0"));
   const [isOnBasket, setIsOnBasket] = useState(false);
-  
   const [currentBundleId, setCurrentBundleId] = useLocalStorage<string>("bundleUuid", "");
 
   const { data: processStatus, startRecoveryProcess, signRecoveryTransactions, blockCountdown } = useRecoveryProcess();
 
+  const startSigning = () => {
+    signRecoveryTransactions(hackedAddress, unsignedTxs, currentBundleId, false);
+  };
   const startRecovery = () => {
     startRecoveryProcess({
       safeAddress,
-      modifyBundleId: val => setCurrentBundleId(val),
+      modifyBundleId: setCurrentBundleId,
       totalGas: totalGasEstimate,
       hackedAddress,
       currentBundleId,
       transactions: unsignedTxs,
     });
   };
-  const getLayoutActiveStep = () => {
+  const getActiveStep = () => {
     if (!!isOnBasket) {
-      return 2;
+      return BundlingSteps.ASSET_SELECTION;
     }
-    if(processStatus !== RecoveryProcessStatus.initial){
-      return 4;
+    if (processStatus !== RecoveryProcessStatus.INITIAL) {
+      return BundlingSteps.SIGN_RECOVERY_TXS;
     }
     if (unsignedTxs.length > 0) {
-      return 3;
+      return BundlingSteps.TX_BUNDLE;
     }
     if (hackedAddress !== "") {
-      return 2;
+      return BundlingSteps.ASSET_SELECTION;
     }
-   
-    return 1;
+    return BundlingSteps.HACKED_ADDRESS_INPUT;
   };
-  const activeStep = getLayoutActiveStep();
+  const reload = () => {
+    localStorage.clear();
+    window.location.reload();
+  };
 
   return (
     <>
       <MetaHeader />
-      {!isConnected && <CustomHeader />}
+
+      <CustomHeader isVisible={!walletConnected} />
+
       <div
         style={{
           display: "flex",
@@ -73,56 +72,38 @@ const Home: NextPage = () => {
           alignItems: "center",
         }}
       >
-        <ConnectStep
-          isVisible={!isConnected}
+        <ConnectionProcess
+          isVisible={!walletConnected}
           safeAddress={safeAddress}
-          setSafeAddress={newAdd => setSafeAddress(newAdd)}
-          address={address ?? ""}
+          setSafeAddress={setSafeAddress}
+          connectedAddress={connectedAddress}
         />
-        {!!isConnected && (
-          <Layout stepActive={activeStep} hackedAddress={hackedAddress} safeAddress={safeAddress}>
-            <>
-              <HackedAddressStep
-                isVisible={activeStep === 1}
-                onSubmit={address => {
-                  setHackedAddress(address);
-                }}
-              />
-              <AssetSelectionStep
-                isVisible={activeStep === 2}
-                onSubmit={txsToAdd => {
-                  setUnsignedTxs(txsToAdd);
-                  setIsOnBasket(false);
-                }}
-                hackedAddress={hackedAddress}
-                safeAddress={safeAddress}
-              />
-              <TransactionBundleStep
-                setTotalGasEstimate={setTotalGasEstimate}
-                totalGasEstimate={totalGasEstimate}
-                isVisible={activeStep === 3}
-                transactions={unsignedTxs}
-                modifyTransactions={setUnsignedTxs}
-                onAddMore={() => setIsOnBasket(true)}
-                clear={() => setUnsignedTxs([])}
-                onSubmit={() => startRecovery()}
-              />
-            </>
-          </Layout>
-        )}
+
+        <BundlingProcess
+          isVisible={walletConnected}
+          activeStep={getActiveStep()}
+          hackedAddress={hackedAddress}
+          safeAddress={safeAddress}
+          totalGasEstimate={totalGasEstimate}
+          unsignedTxs={unsignedTxs}
+          setHackedAddress={setHackedAddress}
+          setUnsignedTxs={setUnsignedTxs}
+          setIsOnBasket={setIsOnBasket}
+          setTotalGasEstimate={setTotalGasEstimate}
+          startRecovery={startRecovery}
+        />
+
         <RecoveryProcess
           recoveryStatus={processStatus}
-          finishProcess={() => {
-            localStorage.clear();
-            window.location.reload();
-          }}
-          startSigning={() => signRecoveryTransactions(hackedAddress, unsignedTxs, currentBundleId, false)}
-          startProcess={() => startRecovery()}
+          finishProcess={reload}
+          startSigning={startSigning}
+          startProcess={startRecovery}
           blockCountdown={blockCountdown}
-          connectedAddress={address ?? ""}
+          connectedAddress={connectedAddress}
           safeAddress={safeAddress}
           hackedAddress={hackedAddress}
-        ></RecoveryProcess>
+        />
+
         {/* <CustomPortal
           title={"Clear cache"}
           description={
